# Speed up `Quantity.__iter__()`